v0.3 -- Teensy Improvments
Teensy state/variable scheme is improved dramatically from csDiscrim and pyStim schemes.

* Teensy still keeps a list of valid variable values and headers of those which can be changed via serial, but it has a way to poll the change.

to change state to S1, for example, you would send a1>, where a is the header flag for the state variable change, 1 is the new value, and the > character is the end flag.

now you can send a<, which would return echo,0,1 meaning variable 0 in the array is set to 1.

Beyond this, there is now a state change block latch and no more assign variable function. We just take the readability hit and reference all variables as their array entry.

These changes are more obvious by reading the documentation and the code.