v0.8 -- substantial improvments
a) Fixed an anoying bug that was causing intermitent timing jitter. The jitter wasn't bad (<0.02% of the samples; at 1-10 ms), but no reason it shouldn't be clock like. I've now benchmarked it on a mediocre mac and I get 0 jitter for up to 500 trials at 1k. Teensy 3.6 set to 180 MHz (default).

What was happening is that when I set the syncState bit to 0 in python, it would keep resending the state change and poluting the serial buffer. Now, I incrment a counter and if python and teensy are out of sync (in task) for longer than 500 ms, it will resend the state change request. I have yet to see it actually resend. I just added that assuming slower computers might need it. The only downside is the state headers now have to reset this variable, not a big deal, just remember for new states.

b) added a running dprime estimate to the outcome plot. It plots itself as a black line.

c) a vector of stimulus choices, non-stimulus (catch) choices, the indicies of the stim trials and non-stim trials are now logged in the HDF as attributes for easy psychometic analysis later.

c) teensy detection code has changed substantially. you will need to reflash the teensy with this.